Move attribute registration to separate file
